Split period form out into separate files
