+++ src/components/period_form/index.js
@@ -1,0 +1,3 @@
+"use strict";
+
+module.exports = require('./index.jsx')
