+++ src/components/period_form/index.jsx
@@ -2,217 +2,14 @@
 
 var React = require('react')
   , Immutable = require('immutable')
-  , TerminusInput
-  , LabelInput
-
-const emptyTerminus = Immutable.fromJS({ label: '', in: { year: '' }})
-
-TerminusInput = React.createClass({
-  propTypes: {
-    terminusType: React.PropTypes.oneOf(['start', 'stop']).isRequired,
-  },
-
-  getDefaultProps: function () {
-    return { terminus: emptyTerminus }
-  },
-
-  getInitialState: function () {
-    return {
-      terminus: this.props.terminus
-    }
-  },
-
-  isMultivalued: function () {
-    return this.state.terminus.hasIn(['in', 'earliestYear']);
-  },
-
-  handleChangeLabel: function (e) {
-    var parseDate = require('../utils/date_parser.js')
-      , parsed = parseDate(e.target.value)
-
-    if (parsed) {
-      this.setState({ terminus: Immutable.fromJS(parsed).delete('_type') })
-    } else {
-      this.setState({ terminus: emptyTerminus })
-    }
-  },
-
-  render: function () {
-    var Input = require('./shared/input.jsx')
-
-    return (
-      <div className="row">
-        <div className="col-md-12">
-          <Input
-              id={`js-${this.props.terminusType}Date`}
-              name="label"
-              label="Label"
-              value={this.state.terminus.get('label')}
-              onChange={this.props.autoparse ? this.handleChangeLabel : this.handleChange} />
-        </div>
-
-        <div>
-          {this.isMultivalued() ?
-            (
-            <div>
-              <div className="col-md-4">
-                <Input
-                    id={`js-${this.props.terminusType}-earliest-start`}
-                    name="earliestStart"
-                    label="Earliest start"
-                    value={this.state.terminus.getIn(['in', 'earliestYear'])}
-                    disabled={this.props.autoparse}
-                    onChange={this.handleChange} />
-              </div>
-              <div className="col-md-4">
-                <Input
-                    id={`js-${this.props.terminusType}-latest-stop`}
-                    name="latestStop"
-                    label="Latest stop"
-                    value={this.state.terminus.getIn(['in', 'latestYear'])}
-                    disabled={this.props.autoparse}
-                    onChange={this.handleChange} />
-              </div>
-            </div>
-            )
-              :
-            (
-            <div>
-              <div className="col-md-4">
-                <Input
-                    id={`js-${this.props.terminusType}Year`}
-                    name="year"
-                    label="Year"
-                    value={this.state.terminus.getIn(['in', 'year'])}
-                    disabled={this.props.autoparse}
-                    onChange={this.handleChange} />
-              </div>
-              <div className="col-md-4">
-              </div>
-            </div>
-            )
-          }
-          <div className="col-md-4">
-            <label> </label>
-            <button className="btn btn-primary" disabled={this.props.autoparse}>
-              Toggle earliest/latest
-            </button>
-          </div>
-        </div>
-
-      </div>
-    )
-  }
-});
-
-var LanguageSelectDropdown = React.createClass({
-  displayName: 'LanguageSelectDropdown',
-  handleSelect: function (language) {
-    this.props.onChange(language);
-  },
-  getMatchingLanguages: function (text) {
-    var languages = require('../utils/languages')
-
-    if (text) {
-      let matchingSet = Immutable.Set(text);
-      languages = languages
-        .filter(lang => matchingSet.subtract(Immutable.Set(lang.get('name'))).size === 0)
-    } else {
-      languages = languages
-        .filter(lang => lang.get('iso6391'))
-        .sortBy(lang => lang.get('name'))
-    }
-
-    return languages.take(10);
-  },
-  render: function () {
-    var DropdownAutocomplete = require('./shared/dropdown_autocomplete.jsx')
-
-    return <DropdownAutocomplete
-      label={this.props.language}
-      getMatchingItems={this.getMatchingLanguages}
-      renderMatch={lang => lang.get('name')}
-      onSelect={this.handleSelect} />
-  }
-});
-
-var ScriptSelectDropdown = React.createClass({
-  displayName: 'ScriptSelectDropdown',
-  handleSelect: function (script) {
-    this.props.onChange(script);
-  },
-  getMatchingScripts: function (text) {
-    var scripts = require('../utils/scripts');
-    if (text) {
-      let matchingSet = Immutable.Set(text);
-      scripts = scripts
-        .filter(script => matchingSet.subtract(Immutable.Set(script.get('name'))).size === 0)
-    }
-
-    return scripts.take(10);
-  },
-  render: function () {
-    var DropdownAutocomplete = require('./shared/dropdown_autocomplete.jsx')
-
-    return <DropdownAutocomplete
-      label={this.props.script}
-      getMatchingItems={this.getMatchingScripts}
-      renderMatch={script => script.get('name')}
-      onSelect={this.handleSelect} />
-  }
-});
-
-LabelInput = React.createClass({
-  handleSelect: function (field, value) {
-    var newLabel = this.props.label.set(field, value.get('code').toLowerCase());
-    this.props.onChange(newLabel);
-  },
-  handleChange: function (field, e) {
-    var newLabel = this.props.label.set(field, e.target.value);
-    this.props.onChange(newLabel);
-  },
-  render: function () {
-    return (
-      <div className="form-group" data-field="alternate-label">
-        <div className="input-group">
-
-          <div className="input-group-btn">
-            <LanguageSelectDropdown
-                onChange={this.handleSelect.bind(null, 'language')}
-                language={this.props.label.get('language')} />
-            <ScriptSelectDropdown
-                onChange={this.handleSelect.bind(null, 'script')}
-                script={this.props.label.get('script')} />
-          </div>
-
-          <input
-              className="form-control"
-              id={this.props.id}
-              value={this.props.label.get('value')}
-              onChange={this.handleChange.bind(null, 'value')}
-              type="text" />
-          {
-            !this.props.handleAddLabel ? null :
-              <span className="input-group-addon btn" onClick={this.props.handleAddLabel}>
-                <strong>+</strong>
-              </span>
-          }
-          {
-            !this.props.handleRemoveLabel ? null :
-              <span className="input-group-addon btn" onClick={this.props.handleRemoveLabel}>
-                <strong>-</strong>
-              </span>
-          }
-        </div>
-      </div>
-    )
-  }
-});
+  , TerminusInput = require('./terminus_input.jsx')
+  , LocalizedLabelInput = require('./localized_label_input.jsx')
+  , SpatialCoverageForm = require('./spatial_coverage_form.jsx')
 
 module.exports = React.createClass({
   getInitialState: function () {
-    var { wasAutoparsed } = require('../helpers/terminus')
-      , { getAlternateLabels } = require('../helpers/period.js')
+    var { wasAutoparsed } = require('../../helpers/terminus')
+      , { getAlternateLabels } = require('../../helpers/period.js')
       , alternateLabels
       , parseDates
 
@@ -265,11 +62,17 @@ module.exports = React.createClass({
   toggleAutoparse: function () {
     this.setState(prev => ({ parseDates: !prev.parseDates }));
   },
+
+  /* *
+   * Change handlers for the various form components
+   * */
+
   handleChange: function (field, e) {
     var value = e.target.value;
     if (!Array.isArray(field)) field = [field];
     this.setState(prev => ({ period: prev.period.setIn(field, value) }));
   },
+
   handleLabelChange: function (label) {
     this.setState(prev => {
       var period = prev.period;
@@ -283,10 +86,12 @@ module.exports = React.createClass({
       return { period }
     });
   },
+
   handleAlternateLabelChange: function (idx, label) {
     var prevAltLabels = this.state.alternateLabels;
     this.setState({ alternateLabels: prevAltLabels.set(idx, label) });
   },
+
   addAlternateLabel: function (i) {
     if (!this.state.alternateLabels.getIn([i, 'value'])) {
       return
@@ -301,6 +106,7 @@ module.exports = React.createClass({
       });
     }
   },
+
   removeAlternateLabel: function (i) {
     var prevAltLabels = this.state.alternateLabels
 
@@ -314,16 +120,17 @@ module.exports = React.createClass({
       });
     }
   },
+
   render: function () {
-    var Input = require('./shared/input.jsx')
-      , { getOriginalLabel } = require('../helpers/period.js')
+    var Input = require('../shared/input.jsx')
+      , { getOriginalLabel } = require('../../helpers/period.js')
 
     return (
       <div className="period-form-body">
         <div className="row">
           <div className="col-md-6 period-form-panel">
             <label className="field-required-label" htmlFor="js-label">Label</label>
-            <LabelInput
+            <LocalizedLabelInput
                 id="js-label"
                 label={getOriginalLabel(this.state.period)}
                 onChange={this.handleLabelChange} />
@@ -331,7 +138,7 @@ module.exports = React.createClass({
             <label htmlFor="js-label">Alternate labels</label>
             {
               this.state.alternateLabels.map((label, i) =>
-                <LabelInput id="js-label"
+                <LocalizedLabelInput id="js-label"
                     key={i}
                     label={label}
                     onChange={this.handleAlternateLabelChange.bind(null, i)}
@@ -363,6 +170,7 @@ module.exports = React.createClass({
         <div className="row">
           <div className="period-form-panel col-md-6">
             <h3>Spatial coverage</h3>
+            <SpatialCoverageForm />
           </div>
           <div className="period-form-panel col-md-6">
             <h3>Temporal coverage</h3>
