+++ src/components/period_form/spatial_coverage_form.jsx
@@ -1,0 +1,14 @@
+"use strict";
+
+var React = require('react')
+
+module.exports = React.createClass({
+  displayName: 'PeriodSpatialCoverageForm',
+  render: function () {
+    return (
+      <div>
+      Spatial coverage
+      </div>
+    )
+  }
+});
